% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Chymo_high}
\alias{Chymo_high}
\title{Product concentration of 101 observed time}
\format{A data frame with 101 rows and 2 variables}
\usage{
Chymo_high
}
\description{
A synthetic data set of the product concentration of the enzyme kinetics
with lower initial enzyme concentration and discrete observed time.
}
\keyword{datasets}
